cleaning up more tests and references
