+++ app/src/internal/java/com/duckduckgo/app/dev/settings/DevSettingsViewModel.kt
@@ -25,7 +25,7 @@ import com.duckduckgo.app.dev.settings.db.UAOverride
 import com.duckduckgo.app.global.DispatcherProvider
 import com.duckduckgo.app.traces.api.StartupTraces
 import com.duckduckgo.di.scopes.ActivityScope
-import com.duckduckgo.savedsites.SavedSitesRepository
+import com.duckduckgo.savedsites.api.SavedSitesRepository
 import javax.inject.Inject
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
