+++ app/src/main/java/com/duckduckgo/app/bookmarks/ui/BookmarksAdapter.kt
@@ -26,7 +26,6 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.duckduckgo.app.bookmarks.model.SavedSite
 import com.duckduckgo.app.browser.R
 import com.duckduckgo.app.browser.databinding.ViewSavedSiteEmptyHintBinding
 import com.duckduckgo.app.browser.favicon.FaviconManager
@@ -34,6 +33,7 @@ import com.duckduckgo.app.global.DispatcherProvider
 import com.duckduckgo.app.global.baseHost
 import com.duckduckgo.mobile.android.databinding.RowTwoLineItemBinding
 import com.duckduckgo.mobile.android.ui.menu.PopupMenu
+import com.duckduckgo.savedsites.api.models.SavedSite
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
