+++ app/src/main/java/com/duckduckgo/app/bookmarks/ui/BookmarksEntityQueryListener.kt
@@ -17,10 +17,10 @@
 package com.duckduckgo.app.bookmarks.ui
 
 import androidx.lifecycle.viewModelScope
-import com.duckduckgo.app.bookmarks.model.BookmarkFolder
-import com.duckduckgo.app.bookmarks.model.SavedSite
 import com.duckduckgo.app.bookmarks.ui.bookmarkfolders.BookmarkFoldersAdapter
 import com.duckduckgo.app.utils.ConflatedJob
+import com.duckduckgo.savedsites.api.models.BookmarkFolder
+import com.duckduckgo.savedsites.api.models.SavedSite
 import java.util.*
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
