+++ app/src/main/java/com/duckduckgo/app/browser/favicon/DuckDuckGoFaviconManager.kt
@@ -22,8 +22,6 @@ import android.graphics.drawable.Drawable
 import android.net.Uri
 import android.widget.ImageView
 import androidx.core.net.toUri
-import com.duckduckgo.app.bookmarks.db.BookmarksDao
-import com.duckduckgo.app.bookmarks.model.SavedSitesRepository
 import com.duckduckgo.app.browser.favicon.FileBasedFaviconPersister.Companion.FAVICON_PERSISTED_DIR
 import com.duckduckgo.app.browser.favicon.FileBasedFaviconPersister.Companion.FAVICON_TEMP_DIR
 import com.duckduckgo.app.browser.favicon.FileBasedFaviconPersister.Companion.NO_SUBFOLDER
@@ -36,12 +34,14 @@ import com.duckduckgo.app.global.view.generateDefaultDrawable
 import com.duckduckgo.app.global.view.loadFavicon
 import com.duckduckgo.app.location.data.LocationPermissionsRepository
 import com.duckduckgo.autofill.api.store.AutofillStore
+import com.duckduckgo.savedsites.api.SavedSitesRepository
+import com.duckduckgo.savedsites.store.SyncEntitiesDao
 import java.io.File
 import kotlinx.coroutines.withContext
 
 class DuckDuckGoFaviconManager constructor(
     private val faviconPersister: FaviconPersister,
-    private val bookmarksDao: BookmarksDao,
+    private val entitiesDao: SyncEntitiesDao,
     private val fireproofWebsiteRepository: FireproofWebsiteRepository,
     private val locationPermissionsRepository: LocationPermissionsRepository,
     private val savedSitesRepository: SavedSitesRepository,
@@ -292,8 +292,7 @@ class DuckDuckGoFaviconManager constructor(
         val query = "%$domain%"
 
         return withContext(dispatcherProvider.io()) {
-            // TODO:
-            bookmarksDao.bookmarksCountByUrl(query) +
+            entitiesDao.countEntitiesByUrl(query) +
                 locationPermissionsRepository.permissionEntitiesCountByDomain(query) +
                 fireproofWebsiteRepository.fireproofWebsitesCountByDomain(domain) +
                 savedSitesRepository.getFavoritesCountByDomain(query) +
