+++ app/src/main/java/com/duckduckgo/app/global/store/AndroidUserBrowserProperties.kt
@@ -16,7 +16,6 @@
 
 package com.duckduckgo.app.global.store
 
-import com.duckduckgo.app.bookmarks.model.SavedSitesRepository
 import com.duckduckgo.app.email.EmailManager
 import com.duckduckgo.app.global.install.AppInstallStore
 import com.duckduckgo.app.global.install.daysInstalled
@@ -26,6 +25,7 @@ import com.duckduckgo.app.widget.ui.WidgetCapabilities
 import com.duckduckgo.browser.api.UserBrowserProperties
 import com.duckduckgo.mobile.android.ui.DuckDuckGoTheme
 import com.duckduckgo.mobile.android.ui.store.ThemingDataStore
+import com.duckduckgo.savedsites.api.SavedSitesRepository
 import java.util.*
 
 class AndroidUserBrowserProperties(
