+++ app/src/main/java/com/duckduckgo/app/pixels/AppPixelName.kt
@@ -237,11 +237,6 @@ enum class AppPixelName(override val pixelName: String) : Pixel.PixelName {
     EMAIL_DID_PRESS_WAITLIST_DIALOG_NOTIFY_ME("email_did_press_waitlist_dialog_notify_me"),
     EMAIL_DID_PRESS_WAITLIST_DIALOG_NO_THANKS("email_did_press_waitlist_dialog_dismiss"),
 
-    BOOKMARK_IMPORT_SUCCESS("m_bi_s"),
-    BOOKMARK_IMPORT_ERROR("m_bi_e"),
-    BOOKMARK_EXPORT_SUCCESS("m_be_a"),
-    BOOKMARK_EXPORT_ERROR("m_be_e"),
-
     ENCRYPTED_IO_EXCEPTION("m_e_io_e"),
     ENCRYPTED_GENERAL_EXCEPTION("m_e_g_e"),
