+++ app/src/main/java/com/duckduckgo/app/widget/FavoritesObserver.kt
@@ -20,10 +20,10 @@ import android.appwidget.AppWidgetManager
 import android.content.ComponentName
 import android.content.Context
 import androidx.lifecycle.LifecycleOwner
-import com.duckduckgo.app.bookmarks.model.SavedSitesRepository
 import com.duckduckgo.app.browser.R
 import com.duckduckgo.app.lifecycle.MainProcessLifecycleObserver
 import com.duckduckgo.di.scopes.AppScope
+import com.duckduckgo.savedsites.api.SavedSitesRepository
 import com.duckduckgo.widget.SearchAndFavoritesWidget
 import dagger.SingleInstanceIn
 import javax.inject.Inject
