+++ saved-sites/saved-sites-api/src/main/java/com/duckduckgo/savedsites/api/models/Models.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.savedsites
+package com.duckduckgo.savedsites.api.models
 
-import com.duckduckgo.savedsites.SavedSite.Bookmark
-import com.duckduckgo.savedsites.SavedSite.Favorite
+import com.duckduckgo.savedsites.api.models.SavedSite.Bookmark
+import com.duckduckgo.savedsites.api.models.SavedSite.Favorite
 import com.duckduckgo.savedsites.store.Entity
 import com.duckduckgo.savedsites.store.Relation
 import java.io.Serializable
