+++ saved-sites/saved-sites-impl/build.gradle
@@ -24,34 +24,34 @@ apply from: "$rootProject.projectDir/gradle/android-library.gradle"
 
 dependencies {
     implementation project(path: ':di')
+    implementation project(path: ':statistics')
     implementation project(path: ':common')
     implementation project(path: ':common-ui')
     implementation project(path: ':saved-sites-api')
-    implementation project(path: ':saved-sites-store')
-
-    implementation "io.reactivex.rxjava2:rxjava:_"
-    implementation "io.reactivex.rxjava2:rxandroid:_"
+    api project(path: ':saved-sites-store')
 
     anvil project(path: ':anvil-compiler')
     implementation project(path: ':anvil-annotations')
 
+    implementation "io.reactivex.rxjava2:rxjava:_"
+    implementation "io.reactivex.rxjava2:rxandroid:_"
     implementation JakeWharton.timber
-
+    implementation Google.dagger
     implementation KotlinX.coroutines.core
     implementation AndroidX.core.ktx
+    implementation 'org.jsoup:jsoup:_'
 
     // Testing dependencies
+    testImplementation project(path: ':common-test')
+    testImplementation AndroidX.core
+    testImplementation AndroidX.archCore.testing
     testImplementation "org.mockito.kotlin:mockito-kotlin:_"
     testImplementation Testing.junit4
-    testImplementation AndroidX.core
     testImplementation AndroidX.test.ext.junit
     testImplementation "androidx.test:runner:_"
     testImplementation Testing.robolectric
     testImplementation CashApp.turbine
     testImplementation "org.jetbrains.kotlin:kotlin-reflect:_"
-
-    testImplementation project(path: ':common-test')
-
     testImplementation (KotlinX.coroutines.test) {
         // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
         // conflicts with mockito due to direct inclusion of byte buddy
