+++ saved-sites/saved-sites-store/src/main/java/com/duckduckgo/savedsites/store/SyncEntitiesDao.kt
@@ -69,6 +69,13 @@ interface SyncEntitiesDao {
     @Query("select * from entities where url = :url limit 1")
     fun entityByUrl(url: String): Entity?
 
+    @Query(
+        "select count(*) from entities where entities.url LIKE :domain",
+    )
+    fun countEntitiesByUrl(
+        domain: String,
+    ): Int
+
     @Query("select CAST(COUNT(*) AS BIT) from entities")
     fun hasEntities(): Boolean
